# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hostel World",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get Events from DB",
                "description": "Get seeded events from database (with valid dates)",
                "operationId": "EventFromDB",
                "parameters": [
                    {
                        "name": "term",
                        "in": "query",
                        "description": "Search by city and country"
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter by exact date"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get array of events",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Event"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "next_page_url": {
                                            "description": "User unique email address",
                                            "type": "string",
                                            "format": "email",
                                            "example": "user@gmail.com"
                                        },
                                        "path": {
                                            "description": "next page url",
                                            "type": "string",
                                            "example": "http://localhost:8001/api/events"
                                        },
                                        "per_page": {
                                            "description": "number of result per page",
                                            "type": "string",
                                            "example": "20"
                                        },
                                        "prev_page_url": {
                                            "description": "previous page url",
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "to": {
                                            "description": "item count",
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "total": {
                                            "description": "total record count",
                                            "type": "string",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/events_json": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get Events From JSON",
                "description": "Get Events from JSON file with intended data intact",
                "operationId": "EventFromJSON",
                "parameters": [
                    {
                        "name": "term",
                        "in": "query",
                        "description": "Search by city and country"
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter by exact date"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get array of events",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Event"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "next_page_url": {
                                            "description": "User unique email address",
                                            "type": "string",
                                            "format": "email",
                                            "example": "user@gmail.com"
                                        },
                                        "path": {
                                            "description": "next page url",
                                            "type": "string",
                                            "example": "http://localhost:8001/api/events"
                                        },
                                        "per_page": {
                                            "description": "number of result per page",
                                            "type": "string",
                                            "example": "20"
                                        },
                                        "prev_page_url": {
                                            "description": "previous page url",
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "to": {
                                            "description": "item count",
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "total": {
                                            "description": "total record count",
                                            "type": "string",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Event": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "John"
                    },
                    "city": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "ikeja"
                    },
                    "country": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Nigeria"
                    },
                    "startDate": {
                        "format": "date-time"
                    },
                    "endDate": {
                        "format": "date-time"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Event"
                }
            },
...
        },
...
        "securitySchemes": {
            "token": {
                "type": "http",
                "name": "token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}